backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.4
exceptiongroup==1.2.1
Faker==25.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.4.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.0
urllib3==2.2.2
Werkzeug==3.0.3
